requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# LLM - Using OpenRouter for all providers
openai>=1.0.0

# Frameworks - MVP versions that work with OpenRouter
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Note: google-adk, openai-agents, orchestral-ai are newer frameworks
# For MVP, we'll use OpenRouter-compatible implementations via openai SDK

# UI
streamlit>=1.40.0

# Async support
asyncio-compat>=0.1.0
